AS when can I commit war crimes
